moved addEntry and submitManualEntry into ManualForm component, changed modal popup to not appear with a successful submit
